backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.4
scipy==1.13.1
python-dotenv==1.0.1
